fix: general error correction for jwt authentication, login, and register #9
- 스웨거 작성 중 에러가 발생하여 처리
- 로그인, 회원가입, 토큰 재발급, 유저정보 기능 구현 및 스웨거 작성

+++ accounts/views.py
@@ -1,38 +1,41 @@
-from django.contrib.auth.hashers import check_password
 from django.core.files.storage import default_storage
 from django.db import IntegrityError
 
 from drf_yasg.utils import swagger_auto_schema
 
 from rest_framework.parsers import MultiPartParser
-from rest_framework.generics import RetrieveUpdateAPIView
-from rest_framework.response import Response
-from rest_framework.exceptions import ValidationError
 from rest_framework.views import APIView
-from rest_framework import request
-from rest_framework import status
 
-from accounts.serializers import UserSerializers, RegisterSerializer, LoginSerializer
-from accounts.models import User
-from core.tokens import TokenResponseSerializer
-from core.tokens import get_user_id
+from accounts.serializers import (
+    UserSerializers,
+    RegisterSerializer,
+    LoginSerializer,
+    TokenRefreshSerializer,
+)
+from core.constants import SystemCodeManager
+from core.responses import Response
+from core.exceptions import raise_exception
+from core.tokens import CustomJWTAuthentication
+from core.responses import Response
 
 
 class RegisterView(APIView):
     serializer_class = RegisterSerializer
 
+    @swagger_auto_schema(
+        operation_id="회원가입",
+        tags=["사용자 인증"],
+        request_body=serializer_class,
+    )
     def post(self, request):
         serializer = self.serializer_class(data=request.data)
 
         if not serializer.is_valid():
-            raise ValidationError("올바른 포멧이 아닙니다.")
-
-        try:
-            serializer.save()
-        except IntegrityError:
-            raise ValidationError(
-                "이미 등록된 이메일 주소입니다.", code=status.HTTP_400_BAD_REQUEST
+            raise_exception(
+                code=SystemCodeManager.get_message("base_code", "INVALID_FORMAT")
             )
+
+        serializer.save()
 
         return Response(data=serializer.data)
 
@@ -40,66 +43,79 @@ class RegisterView(APIView):
 class LoginView(APIView):
     serializer_class = LoginSerializer
 
+    @swagger_auto_schema(
+        operation_id="로그인",
+        tags=["사용자 인증"],
+        request_body=LoginSerializer,
+    )
     def post(self, request):
-        email = request.POST.get("email")
-        password = request.POST.get("password")
+        serializer = self.serializer_class(data=request.data)
 
-        try:
-            user = User.objects.get(email=email)
-
-            if check_password(password, user.password):
-                serializer = TokenResponseSerializer(user)
-                return Response(data=serializer.to_representation(serializer))
-            else:
-                raise ValidationError(
-                    "패스워드가 잘못되었습니다.", code=status.HTTP_400_BAD_REQUEST
-                )
-
-        except User.DoesNotExist:
-            raise ValidationError(
-                "가입되지 않은 사용자입니다.", code=status.HTTP_400_BAD_REQUEST
+        if not serializer.is_valid():
+            raise_exception(
+                code=SystemCodeManager.get_message("base_code", "INVALID_FORMAT")
             )
 
+        return Response(data=serializer.data)
 
-class UserInfoViews(RetrieveUpdateAPIView):
+
+class TokenRefreshView(APIView):
+    serializer_class = TokenRefreshSerializer
+
+    @swagger_auto_schema(
+        operation_id="토큰 재발급",
+        tags=["사용자 인증"],
+        request_body=TokenRefreshSerializer,
+    )
+    def post(self, request):
+        """
+        토큰 재발급을 처리합니다.
+        """
+        serializer = self.serializer_class(data=request.data)
+
+        # Validation Check
+        if not serializer.is_valid():
+            raise_exception(
+                code=SystemCodeManager.get_message("base_code", "INVALID_FORMAT")
+            )
+
+        return Response(data=serializer.data)
+
+
+class UserInfoViews(APIView):
     parser_classes = (MultiPartParser,)
     serializer_class = UserSerializers
-    http_method_names = ["get", "patch"]
 
     def get_object(self):
-        user = get_user_id(self.request)
+        user = CustomJWTAuthentication().authenticate(self.request)
         return user
 
     @swagger_auto_schema(tags=["유저 정보"])
     def get(self, request, *args, **kwargs):
         """
         유저 정보 조회
-        ---
         """
-        return super().get(request, *args, **kwargs)
+        instance = self.get_object()
+        serializer = self.serializer_class(instance)
+        return Response(data=serializer.data)
 
     @swagger_auto_schema(tags=["유저 정보"])
     def patch(self, request, *args, **kwargs):
         """
         유저 정보 부분 수정
-        ---
         """
-        return super().patch(request, *args, **kwargs)
-
-    def retrieve(self, request, *args, **kwargs):
-        instance = self.get_object()
-        serializer = self.serializer_class(instance)
-        return Response(serializer.data)
-
-    def update(self, request, *args, **kwargs):
         instance = self.get_object()
         serializer = self.serializer_class(instance, data=request.data, partial=True)
-        serializer.is_valid(raise_exception=True)
-        self.perform_update(serializer)
-        return Response(serializer.data, status=status.HTTP_200_OK)
+        if serializer.is_valid(raise_exception=True):
+            self.perform_update(serializer)
+            return Response(data=serializer.data)
+        raise_exception(code=SystemCodeManager.get_message("base_code", "BAD_REQUEST"))
 
     def perform_update(self, serializer):
         instance = self.get_object()
-        if instance.profile_img.path != "img/default/default_img.jpg":
+        if (
+            hasattr(instance, "profile_img")
+            and instance.profile_img.path != "img/default/default_img.jpg"
+        ):
             default_storage.delete(instance.profile_img.path)
         serializer.save()
